requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.9.0
appdirs==1.4.4
argbind==0.3.9
asttokens==3.0.0
async-generator==1.10
async-timeout==5.0.1
attrs==22.1.0
audioread==3.0.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.22
botocore==1.36.22
cachetools==5.3.2
certifi==2025.1.31
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
chess==1.11.2
click==8.1.3
colorama==0.4.5
comtypes==1.2.0
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
decorator==4.4.2
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
distlib==0.3.8
distro==1.9.0
docstring_parser==0.16
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.2.0
executing==2.2.0
fake-useragent==1.4.0
fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
flatten-dict==0.4.2
flet==0.27.6
flet-desktop==0.27.6
flet-web==0.27.6
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.25.0
google-genai==1.7.0
google-generativeai==0.8.5
googleapis-common-protos==1.62.0
gradio==5.27.0
gradio_client==1.9.0
greenlet==3.0.3
groovy==0.1.2
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huepy==1.2.1
huggingface-hub==0.30.2
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.4.7
importlib_resources==6.5.2
iniconfig==2.0.0
instabot==0.117.0
instaloader==4.14.1
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.2.0
more-itertools==10.6.0
MouseInfo==0.1.3
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
nari-tts @ git+https://github.com/nari-labs/dia.git@48da80e0f5d06f95c2c4f00f45228ab47c8c01ca
networkx==3.2.1
newsapi-python==0.2.7
newsdataapi==0.1.21
newspaper3k==0.2.8
nltk==3.8.1
numba==0.61.2
numpy==2.2.5
oauthlib==3.2.2
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.11.0.86
orjson==3.10.16
outcome==1.2.0
packaging==23.1
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.2.3
parse==1.20.1
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.0
playwright==1.42.0
pluggy==1.3.0
pooch==1.8.2
proglog==0.1.10
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.23.0
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyee==8.2.2
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyhub==0.0.11
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyloudnorm==0.1.1
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pystoi==0.4.1
pytesseract==0.3.13
pytest==7.4.1
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.90
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
pytweening==1.0.4
pytz==2025.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
randomname==0.2.1
regex==2024.11.6
repath==0.9.0
requests==2.32.3
requests-auth-aws-sigv4==0.7
requests-file==2.1.0
requests-toolbelt==1.0.0
responses==0.25.7
rich==14.0.0
rsa==4.9
ruff==0.11.7
s3transfer==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scapy==2.6.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
selenium==4.29.0
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
tiktok-uploader==1.0.16
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
torch==2.6.0
torch-stoi==0.2.3
torchaudio==2.6.0
torchvision==0.20.1+cu121
tqdm==4.64.1
traitlets==5.14.3
transformers==4.48.0
trio==0.22.0
trio-websocket==0.9.2
triton-windows==3.2.0.post18
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==1.26.20
ursina==7.0.0
uvicorn==0.34.2
virtualenv==20.25.0
w3lib==2.3.1
watchfiles==1.0.5
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
wsproto==1.2.0
yarl==1.18.3
yt-dlp==2025.1.26
zipp==3.17.0
